simplified mounter and registrar configuration
