+++ src/registrar/index.js
@@ -24,47 +24,129 @@
 
 const path = require('path')
 
-const Registrar = require('./registrar')
+/**
+ * The available {@link Registrar} implementation constructors mapped against their name.
+ *
+ * @private
+ * @type {Map.<string, Function>}
+ */
+const types = new Map()
 
 /**
- * An implementation of {@link Registrar} where the index files in each directory is loaded as a module and the result
- * should be an object and containing properties whose names are mapped to supported verbs and whose values are one or
- * more handlers which are then mounted against that verb.
+ * Responsible for loading routes from a specific structure of modules and mounting them onto the server using a given
+ * {@link Mounter}.
  *
  * @public
- * @extends Registrar
  */
-class IndexRegistrar extends Registrar {
+class Registrar {
 
   /**
-   * @inheritDoc
-   * @override
+   * Defines the specified <code>type</code> of {@link Registrar} so that it can be looked up using its name.
+   *
+   * @param {Function} type - the constructor of the {@link Registrar} implementation to be defined
+   * @return {Function} A reference to <code>type</code>.
+   * @public
+   * @static
    */
-  static name() {
-    return 'index'
+  static define(type) {
+    types.set(type.getName(), type)
+
+    return type
   }
 
   /**
-   * @inheritDoc
-   * @override
+   * Returns the name of the {@link Registrar} which can be used to lookup constructors.
+   *
+   * @return {string} The name.
+   * @public
+   * @static
+   * @abstract
+   */
+  static getName() {
+  }
+
+  /**
+   * Looks up the type of {@link Registrar} associated with the specified <code>name</code>.
+   *
+   * @param {string} name - the name associated with the {@link Registrar} implementation to be looked up
+   * @return {Function} The constructor of the {@link Registrar} implementation associated with <code>name</code>.
+   * @public
+   * @static
+   */
+  static lookup(name) {
+    return types.get(name)
+  }
+
+  /**
+   * Creates an instance of {@link Registrar} with the specified <code>mounter</code>.
+   *
+   * @param {Mounter} mounter - the {@link Mounter} to be used
+   * @public
+   */
+  constructor(mounter) {
+    /**
+     * The {@link Mounter} to be used by this {@link Registrar} to mount discovered routes onto the server.
+     *
+     * @protected
+     * @type {Mounter}
+     */
+    this.mounter = mounter
+  }
+
+  /**
+   * Builds the route URL from the <code>file</code> path provided.
+   *
+   * This includes detecting files/directories that represent parameter path variables and inserting the names into the
+   * route URL correctly based on the {@link Mounter}.
+   *
+   * By default, this method simply concatenates the <code>file</code> path segments into the URL while dropping the
+   * last path segment.
+   *
+   * @param {string} file - the file path of the module from which the routes are being loaded
+   * @param {routerify~options} options - the options to be used
+   * @return {string} The route URL built from the <code>file</code> path.
+   * @protected
+   */
+  buildUrl(file, options) {
+    return file.split(path.sep)
+      .slice(0, -1)
+      .reduce((memo, segment) => {
+        const match = segment.match(options.paramPattern)
+        if (match) {
+          segment = this.mounter.formatParamPath(match[1])
+        }
+
+        return `${memo}/${segment}`
+      }, '')
+  }
+
+  /**
+   * Loads the router(s) from the module at the specified <code>file</code> path.
+   *
+   * @param {string} file - the file path of the module from which the routes are being loaded
+   * @param {routerify~options} options - the options to be used
+   * @return {*} The router loaded from the module at the <code>file</code> path.
+   * @protected
+   */
+  loadRouter(file, options) {
+    return require(path.resolve(options.dir, file))
+  }
+
+  /**
+   * Loads routes from the module at the specified <code>file</code> path and then mounts them onto the server.
+   *
+   * @param {string} file - the file path of the module from which the routes are to be loaded
+   * @param {routerify~options} options - the options to be used
+   * @return {void}
+   * @public
+   * @abstract
    */
   register(file, options) {
-    const name = path.basename(file, options.ext)
-    if (name !== 'index') {
-      return
-    }
-
-    const router = this.loadRouter(file, options)
-    const url = this.buildUrl(file, options)
-
-    options.verbs.forEach((verb) => {
-      const handlers = router[verb]
-      if (handlers) {
-        this.mounter.mount(url, verb, handlers, options)
-      }
-    })
   }
 
 }
 
-module.exports = IndexRegistrar
+module.exports = Registrar
+
+require('./index-registrar')
+require('./verb-registrar')
